updates mapping domain package
