+++ mapping/mapping-key-value/src/main/java/jakarta/nosql/mapping/kv/KeyValueTemplateProducer.java
@@ -13,18 +13,23 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
-package jakarta.nosql.mapping.key;
+package jakarta.nosql.mapping.kv;
 
 
-import jakarta.nosql.kv.KeyValueEntity;
+import jakarta.nosql.kv.BucketManager;
 
-public interface KeyValueEntityPrePersist {
+/**
+ * The producer of {@link KeyValueTemplate}
+ */
+public interface KeyValueTemplateProducer {
 
     /**
-     * The {@link jakarta.nosql.kv.KeyValueEntity}  after be saved
+     * creates a {@link KeyValueTemplate}
      *
-     * @return the {@link KeyValueEntity} instance
+     * @param <T>     the KeyValueTemplate instance
+     * @param manager the manager
+     * @return a new instance
+     * @throws NullPointerException when manager is null
      */
-    KeyValueEntity getEntity();
-
+    <T extends KeyValueTemplate> T get(BucketManager manager);
 }
