+++ mapping/mapping-key-value/src/main/java/jakarta/nosql/mapping/kv/KeyValueWorkflow.java
@@ -13,7 +13,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
-package jakarta.nosql.mapping.key;
+package jakarta.nosql.mapping.kv;
 
 
 import jakarta.nosql.kv.KeyValueEntity;
