+++ mapping/mapping-key-value/src/main/java/jakarta/nosql/mapping/kv/package-info.java
@@ -28,4 +28,4 @@
  * Some databases support ordering of keys. There are various hardware implementations, and some users maintain
  * data in memory (RAM), while others employ solid-state drives or rotating disks.
  */
-package jakarta.nosql.mapping.key;
+package jakarta.nosql.mapping.kv;
